Move to `cpython/pyhash.rs`
